+++ extension/src/goGenerateTests.ts
@@ -19,6 +19,7 @@ import { outputChannel } from './goStatus';
 import { getBinPath, resolvePath } from './util';
 import { CommandFactory } from './commands';
 import { GoExtensionContext } from './context';
+import { TelemetryKey, telemetryReporter } from './goTelemetry';
 
 const generatedWord = 'Generated ';
 
@@ -165,6 +166,8 @@ function generateTests(
 	goConfig: vscode.WorkspaceConfiguration
 ): Promise<boolean> {
 	return new Promise<boolean>((resolve, reject) => {
+		telemetryReporter.add(TelemetryKey.TOOL_USAGE_GOTESTS, 1);
+
 		const cmd = getBinPath('gotests');
 		let args = ['-w'];
 		const goGenerateTestsFlags: string[] = goConfig['generateTestsFlags'] || [];
@@ -194,7 +197,7 @@ function generateTests(
 		}
 
 		cp.execFile(cmd, args, { env: toolExecutionEnvironment() }, (err, stdout, stderr) => {
-			outputChannel.appendLine('Generating Tests: ' + cmd + ' ' + args.join(' '));
+			outputChannel.info('Generating Tests: ' + cmd + ' ' + args.join(' '));
 
 			try {
 				if (err && (<any>err).code === 'ENOENT') {
@@ -224,7 +227,7 @@ function generateTests(
 				}
 
 				vscode.window.showInformationMessage(message);
-				outputChannel.append(message);
+				outputChannel.info(message);
 
 				if (testsGenerated && !conf.isTestFile) {
 					toggleTestFile(ctx, goCtx)();
@@ -233,7 +236,7 @@ function generateTests(
 				return resolve(true);
 			} catch (e) {
 				vscode.window.showInformationMessage((e as any).msg);
-				outputChannel.append((e as any).msg);
+				outputChannel.info((e as any).msg);
 				reject(e);
 			}
 		});